remove compiler warnings w/o updating coda-oss

+++ modules/c/nitf/tests/test_imsub_clone.c
@@ -85,7 +85,7 @@ void showFileHeader(nitf_FileHeader* header)
         printf("\tThe length of IMAGE subheader [%d]: %ld bytes\n",
                i, (long)len);
         printf("\tThe length of the IMAGE data: %llu bytes\n\n",
-               dataLen);
+               (long long unsigned int) dataLen);
     }
 
     return;
